// src/backend/src/common/enums/gov-ministries.enum.ts
export enum GovMinistriesEnum {
  ADVANCED_EDUCATION_AND_SKILLS_TRAINING = 'ADVANCED_EDUCATION_AND_SKILLS_TRAINING',
  AGRICULTURE_AND_FOOD = 'AGRICULTURE_AND_FOOD',
  ATTORNEY_GENERAL = 'ATTORNEY_GENERAL',
  BC_PUBLIC_SERVICE_AGENCY = 'BC_PUBLIC_SERVICE_AGENCY',
  CHILDREN_AND_FAMILY_DEVELOPMENT = 'CHILDREN_AND_FAMILY_DEVELOPMENT',
  CITIZENS_SERVICES = 'CITIZENS_SERVICES',
  EDUCATION_AND_CHILD_CARE = 'EDUCATION_AND_CHILD_CARE',
  ENERGY_MINES_AND_LOW_CARBON_INNOVATION = 'ENERGY_MINES_AND_LOW_CARBON_INNOVATION',
  ENVIRONMENT_AND_CLIMATE_CHANGE_STRATEGY = 'ENVIRONMENT_AND_CLIMATE_CHANGE_STRATEGY',
  FINANCE = 'FINANCE',
  FORESTS = 'FORESTS',
  GOVERNMENT_COMMUNICATIONS_AND_PUBLIC_ENGAGEMENT = 'GOVERNMENT_COMMUNICATIONS_AND_PUBLIC_ENGAGEMENT',
  HEALTH = 'HEALTH',
  INDIGENOUS_RELATIONS_AND_RECONCILIATION = 'INDIGENOUS_RELATIONS_AND_RECONCILIATION',
  JOBS_ECONOMIC_RECOVERY_AND_INNOVATION = 'JOBS_ECONOMIC_RECOVERY_AND_INNOVATION',
  LABOUR = 'LABOUR',
  LAND_WATER_AND_RESOURCE_STEWARDSHIP = 'LAND_WATER_AND_RESOURCE_STEWARDSHIP',
  LIQUOR_DISTRIBUTION_BRANCH = 'LIQUOR_DISTRIBUTION_BRANCH',
  MENTAL_HEALTH_AND_ADDICTIONS = 'MENTAL_HEALTH_AND_ADDICTIONS',
  MUNICIPAL_AFFAIRS = 'MUNICIPAL_AFFAIRS',
  OFFICE_OF_THE_PREMIER = 'OFFICE_OF_THE_PREMIER',
  PUBLIC_SAFETY_AND_SOLICITOR_GENERAL_AND_EMERGENCY_BC = 'PUBLIC_SAFETY_AND_SOLICITOR_GENERAL_AND_EMERGENCY_BC',
  SOCIAL_DEVELOPMENT_AND_POVERTY_REDUCTION = 'SOCIAL_DEVELOPMENT_AND_POVERTY_REDUCTION',
  TOURISM_ARTS_CULTURE_AND_SPORT = 'TOURISM_ARTS_CULTURE_AND_SPORT',
  TRANSPORTATION_AND_INFRASTRUCTURE = 'TRANSPORTATION_AND_INFRASTRUCTURE',
}